+++ integration-adnetworks/src/test/java/com/inmobi/adserve/channels/adnetworks/DCPxAdAdNetworkTest.java
@@ -1,29 +1,23 @@
 package com.inmobi.adserve.channels.adnetworks;
 
-import static org.easymock.EasyMock.expect;
-import static org.easymock.classextension.EasyMock.createMock;
-import static org.easymock.classextension.EasyMock.replay;
-
-import java.io.File;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-
+import com.inmobi.adserve.channels.adnetworks.xad.DCPxAdAdNetwork;
+import com.inmobi.adserve.channels.api.*;
+import com.inmobi.adserve.channels.entity.ChannelSegmentEntity;
 import junit.framework.TestCase;
-
 import org.apache.commons.configuration.Configuration;
 import org.jboss.netty.bootstrap.ClientBootstrap;
 import org.jboss.netty.channel.MessageEvent;
 import org.jboss.netty.handler.codec.http.HttpResponseStatus;
 import org.testng.annotations.Test;
 
-import com.inmobi.adserve.channels.adnetworks.xad.DCPxAdAdNetwork;
-import com.inmobi.adserve.channels.api.CasInternalRequestParameters;
-import com.inmobi.adserve.channels.api.Formatter;
-import com.inmobi.adserve.channels.api.HttpRequestHandlerBase;
-import com.inmobi.adserve.channels.api.SASRequestParameters;
-import com.inmobi.adserve.channels.api.SlotSizeMapping;
-import com.inmobi.adserve.channels.entity.ChannelSegmentEntity;
+import java.io.File;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+
+import static org.easymock.EasyMock.expect;
+import static org.easymock.classextension.EasyMock.createMock;
+import static org.easymock.classextension.EasyMock.replay;
 
 
 public class DCPxAdAdNetworkTest extends TestCase {
@@ -73,7 +67,7 @@ public class DCPxAdAdNetworkTest extends TestCase {
         sasParams
                 .setUserAgent("Mozilla%2F5.0+%28iPhone%3B+CPU+iPhone+OS+5_0+like+Mac+OS+X%29+AppleWebKit%2F534.46+%28KHTML%2C+like+Gecko%29+Mobile%2F9A334");
         casInternalRequestParameters.latLong = "37.4429,-122.1514";
-        sasParams.setSlot("9");
+        sasParams.setSlot(Short.valueOf("9"));
         casInternalRequestParameters.uid = "202cb962ac59075b964b07152d234b70";
         String clurl = "http://c2.w.inmobi.com/c.asm/4/b/bx5/yaz/2/b/a5/m/0/0/0/202cb962ac59075b964b07152d234b70/4f8d98e2-4bbd-40bc-87e5-22da170600f9/-1/1/9cddca11?ds=1";
         sasParams.setImpressionId("4f8d98e2-4bbd-40bc-8795-22da170700f9");
@@ -93,7 +87,7 @@ public class DCPxAdAdNetworkTest extends TestCase {
         sasParams
                 .setUserAgent("Mozilla%2F5.0+%28iPhone%3B+CPU+iPhone+OS+5_0+like+Mac+OS+X%29+AppleWebKit%2F534.46+%28KHTML%2C+like+Gecko%29+Mobile%2F9A334");
         casInternalRequestParameters.latLong = "37.4429,-122.1514";
-        sasParams.setSlot("9");
+        sasParams.setSlot(Short.valueOf("9"));
         String clurl = "http://c2.w.inmobi.com/c.asm/4/b/bx5/yaz/2/b/a5/m/0/0/0/202cb962ac59075b964b07152d234b70/4f8d98e2-4bbd-40bc-87e5-22da170600f9/-1/1/9cddca11?ds=1";
         sasParams.setImpressionId("4f8d98e2-4bbd-40bc-8795-22da170700f9");
         String externalKey = "f6wqjq1r5v";
@@ -171,7 +165,7 @@ public class DCPxAdAdNetworkTest extends TestCase {
         casInternalRequestParameters.latLong = "37.4429,-122.1514";
         sasParams.setImpressionId("4f8d98e2-4bbd-40bc-8795-22da170700f9");
         casInternalRequestParameters.uid = "202cb962ac59075b964b07152d234b70";
-        sasParams.setSlot("9");
+        sasParams.setSlot(Short.valueOf("9"));
         sasParams.setOsId(3);
         sasParams.setSiteId("12345");
         String externalKey = "1324";
@@ -202,7 +196,7 @@ public class DCPxAdAdNetworkTest extends TestCase {
         casInternalRequestParameters.latLong = "37.4429,-122.1514";
         sasParams.setImpressionId("4f8d98e2-4bbd-40bc-8795-22da170700f9");
         casInternalRequestParameters.uidIFA = "202cb962ac59075b964b07152d234b70";
-        sasParams.setSlot("9");
+        sasParams.setSlot(Short.valueOf("9"));
         sasParams.setOsId(5);
         sasParams.setSiteId("12345");
         sasParams.setCategories(siteCategories);
@@ -231,7 +225,7 @@ public class DCPxAdAdNetworkTest extends TestCase {
         casInternalRequestParameters.latLong = "37.4429,-122.1514";
         sasParams.setImpressionId("4f8d98e2-4bbd-40bc-8795-22da170700f9");
         casInternalRequestParameters.uidO1 = "202cb962ac59075b964b07152d234b70";
-        sasParams.setSlot("9");
+        sasParams.setSlot(Short.valueOf("9"));
         sasParams.setOsId(5);
         sasParams.setSiteId("12345");
         sasParams.setLocSrc("derived-lat-lon");
@@ -259,7 +253,7 @@ public class DCPxAdAdNetworkTest extends TestCase {
         casInternalRequestParameters.latLong = "";
         sasParams.setImpressionId("4f8d98e2-4bbd-40bc-8795-22da170700f9");
         casInternalRequestParameters.uid = "202cb962ac59075b964b07152d234b70";
-        sasParams.setSlot("15");
+        sasParams.setSlot(Short.valueOf("15"));
         sasParams.setOsId(3);
         sasParams.setCategories(new ArrayList<Long>());
         String externalKey = "1324";
@@ -283,7 +277,7 @@ public class DCPxAdAdNetworkTest extends TestCase {
         casInternalRequestParameters.latLong = "11.6,-11.87";
         sasParams.setCategories(Arrays.asList(new Long[] { 9l, 10l, 13l, 30l }));
         sasParams.setImpressionId("4f8d98e2-4bbd-40bc-8795-22da170700f9");
-        sasParams.setSlot("15");
+        sasParams.setSlot(Short.valueOf("15"));
         sasParams.setOsId(3);
         sasParams.setCategories(new ArrayList<Long>());
         String externalKey = "1324";
@@ -332,7 +326,7 @@ public class DCPxAdAdNetworkTest extends TestCase {
         sasParams.setRemoteHostIp("206.29.182.240");
         sasParams.setUserAgent("Mozilla");
         sasParams.setOsId(3);
-        sasParams.setSlot("4");
+        sasParams.setSlot(Short.valueOf("4"));
         sasParams.setSiteId("1234567");
         String externalKey = "19100";
         String beaconUrl = "http://c2.w.inmobi.com/c.asm/4/b/bx5/yaz/2/b/a5/m/0/0/0"
@@ -358,7 +352,7 @@ public class DCPxAdAdNetworkTest extends TestCase {
         CasInternalRequestParameters casInternalRequestParameters = new CasInternalRequestParameters();
         sasParams.setRemoteHostIp("206.29.182.240");
         sasParams.setUserAgent("Mozilla");
-        sasParams.setSlot("4");
+        sasParams.setSlot(Short.valueOf("4"));
         sasParams.setSource("wap");
         String externalKey = "19100";
         String beaconUrl = "http://c2.w.inmobi.com/c"
@@ -384,7 +378,7 @@ public class DCPxAdAdNetworkTest extends TestCase {
         CasInternalRequestParameters casInternalRequestParameters = new CasInternalRequestParameters();
         sasParams.setRemoteHostIp("206.29.182.240");
         sasParams.setUserAgent("Mozilla");
-        sasParams.setSlot("4");
+        sasParams.setSlot(Short.valueOf("4"));
         String externalKey = "19100";
         sasParams.setSource("app");
         String beaconUrl = "http://c2.w.inmobi.com/c"
@@ -410,7 +404,7 @@ public class DCPxAdAdNetworkTest extends TestCase {
         CasInternalRequestParameters casInternalRequestParameters = new CasInternalRequestParameters();
         sasParams.setRemoteHostIp("206.29.182.240");
         sasParams.setUserAgent("Mozilla");
-        sasParams.setSlot("4");
+        sasParams.setSlot(Short.valueOf("4"));
         String externalKey = "19100";
         sasParams.setSource("app");
         sasParams.setSdkVersion("a370");
